Removed unnecesary type files, destructured mutation object
