--- src/config/apiTypes.ts
@@ -1,7 +1,0 @@
-export type Error = {
-  errors: Record<string, string[]> | string[];
-};
-
-export type Response = {
-  data: Record<string, string | number>;
-};
