+++ src/screens/Signup/index.tsx
@@ -12,10 +12,18 @@ import {
   passwordValidation,
   passwordConfirmationValidation
 } from 'utils/formValidations';
-import { SignUpValues } from 'utils/types';
 import { signUp } from 'services/userService';
 
 import styles from './styles.module.scss';
+
+export interface SignUpValues {
+  email: string;
+  firstName: string;
+  lastName: string;
+  locale: string;
+  password: string;
+  passwordConfirmation: string;
+}
 
 function Signup() {
   const { t } = useTranslation();
@@ -24,18 +32,18 @@ function Signup() {
     register,
     formState: { errors },
     handleSubmit,
-    getValues
+    watch
   } = useForm<SignUpValues>();
 
-  const signupMutation = useMutation((data: SignUpValues) => signUp(data), {
+  const { error, isLoading, mutate, reset } = useMutation((data: SignUpValues) => signUp(data), {
     onSuccess: data => {
       console.log(data);
     }
   });
 
   const onSubmit = handleSubmit(formData => {
-    signupMutation.reset();
-    signupMutation.mutate({ ...formData, locale: 'en' });
+    reset();
+    mutate({ ...formData, locale: 'en' });
   });
 
   return (
@@ -74,18 +82,16 @@ function Signup() {
               name="passwordConfirmation"
               type="password"
               error={errors.passwordConfirmation?.message}
-              inputRef={register(passwordConfirmationValidation(t, getValues('password')))}
+              inputRef={register(passwordConfirmationValidation(t, watch('password')))}
             />
-            {signupMutation.isLoading ? (
+            {isLoading ? (
               <Loading className="self-center" />
             ) : (
               <button className="btn" type="submit">
                 {t('common:signup')}
               </button>
             )}
-            {signupMutation.error && (
-              <span className={`text-error ${styles.submitError}`}>{t('Signup:submitError')}</span>
-            )}
+            {error && <span className={`text-error ${styles.submitMessage}`}>{t('Signup:submitError')}</span>}
           </form>
           <button className="btn secondary" type="button">
             {t('common:login')}
