+++ src/screens/Signup/styles.module.scss
@@ -20,7 +20,7 @@
       margin-bottom: 16px;
       padding-bottom: 16px;
 
-      .submit-error {
+      .submit-message {
         align-self: center;
         margin-top: 8px;
       }
