+++ src/services/userService.ts
@@ -1,5 +1,4 @@
 import api from 'config/api';
-import { Error, Response } from 'config/apiTypes';
-import { SignUpValues } from 'utils/types';
+import { ApiError } from 'utils/types';
 
-export const signUp = (data: SignUpValues) => api.post<Response, Error>('/users', data);
+export const signUp = <T>(data: T) => api.post<Response, ApiError>('/users', data);
