+++ src/utils/types.ts
@@ -1,10 +1,5 @@
 export type Nullable<T> = T | null;
 
-export interface SignUpValues {
-  email: string;
-  firstName: string;
-  lastName: string;
-  locale: string;
-  password: string;
-  passwordConfirmation: string;
+export type ApiError = {
+  errors: Record<string, string[]> | string[];
 };
